restructure payment aggregate
