+++ payment/command-side/src/main/resources/application.yaml
