+++ payment/event-handler/src/main/java/org/example/event/sourcing/order/poc/handler/payment/domain/PaymentEventHandler.java
@@ -1,0 +1,11 @@
+package org.example.event.sourcing.order.poc.handler.payment.domain;
+
+import org.example.event.sourcing.order.poc.common.model.event.PaymentEvent;
+
+import java.util.concurrent.CompletableFuture;
+
+public interface PaymentEventHandler {
+
+    CompletableFuture<Void> onEvent(PaymentEvent event);
+
+}
