+++ payment/query-side/build.gradle
