+++ payment/query-side/src/main/java/org/example/event/sourcing/order/poc/query/payment/QuerySidePaymentApplication.java
