+++ settings.gradle
@@ -1,12 +1,10 @@
 rootProject.name = 'event-sourcing-order-poc'
 
 include 'common'
-include 'command-side-payment'
-include 'event-handler-payment'
-include 'query-side-payment'
 include 'command-side-shipment'
 include 'event-handler-shipment'
 include 'query-side-shipment'
+
 include 'order'
 include 'order:command-side'
 findProject(':order:command-side')?.name = 'command-side'
@@ -14,4 +12,12 @@ include 'order:event-handler'
 findProject(':order:event-handler')?.name = 'event-handler'
 include 'order:query-side'
 findProject(':order:query-side')?.name = 'query-side'
+
+include 'payment'
+include 'payment:command-side'
+findProject(':payment:command-side')?.name = 'command-side'
+include 'payment:event-handler'
+findProject(':payment:event-handler')?.name = 'event-handler'
+include 'payment:query-side'
+findProject(':payment:query-side')?.name = 'query-side'
